Move the SolrLogPost tool into our cli package...
